refactore: move `Application` into separate file
